shallot: sweep character
Rows consumed: character (merge core.ts into index.ts; drive and sweep
stay on the character barrel), player (keep; repoints to ../character).
The character core subpath points at index.ts until the exports pass.
