+++ package.json
@@ -52,7 +52,7 @@
         "./render/core": "./src/standard/render/index.ts",
         "./sear/core": "./src/standard/sear/index.ts",
         "./physics/core": "./src/standard/physics/index.ts",
-        "./character/core": "./src/standard/character/core.ts",
+        "./character/core": "./src/standard/character/index.ts",
         "./bvh/core": "./src/standard/bvh/index.ts",
         "./cells/core": "./src/extras/cells/core.ts",
         "./gltf/core": "./src/extras/gltf/core.ts",
