+++ src/standard/character/index.ts
@@ -306,3 +306,16 @@ export const CharacterPlugin: Plugin = {
         charSig = FNV_BASIS;
     },
 };
+
+// Character extension surface — the eid-keyed drive (`move` / `jump`) + readback (`pose` / `grounded`), for
+// the gym `character` scenario + custom controllers. The happy path (the `Character` component +
+// `CharacterPlugin`, which registers every `[Character, Body]` with the solver) ships on the barrel.
+
+export { grounded, jump, move, pose, teleport } from "./drive";
+export {
+    type CharState,
+    MAX_CHAR_CANDIDATES,
+    type SweepBody,
+    type SweepDiag,
+    sweepCharacter,
+} from "./sweep";
